backend/pyproject.toml
[project]
name = "meeting-summary-backend"
version = "0.1.0"
description = "Backend API for AI-powered meeting summary from audio files"
authors = [
    {name = "AI Agent", email = "ai@agent.com"}
]
readme = "README.md"
requires-python = "3.10.11"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "openai (>=1.52.2,<2.0.0)",
    "pydantic (>=2.10.3,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "bcrypt (==4.0.1)",
    "aiofiles (>=24.1.0,<25.0.0)",
]

[tool.poetry]
packages = [{include = "meeting_summary", from = "src"}]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
